Moved statically defined operations into a single class. Added tests, bug fixes
